Actualización (Desafiate sin terminar)

+++ js/product-info.js
@@ -31,22 +31,43 @@ document.addEventListener('DOMContentLoaded', () => {
 });
 
 
-let arrayProductos = [];
+//Botón que agrega el producto seleccionado a la lista
 document.addEventListener("DOMContentLoaded", () => {
     document.getElementById("comprar").addEventListener("click", () => {
-        let idProduct = localStorage.getItem("IdProduct");
-        let URL = "https://japceibal.github.io/emercado-api/products/" + idProduct + ".json"
-        fetch(URL)
-            .then(res => res.json())
-            .then(data => {
-               arrayProductos= JSON.parse(localStorage.getItem('arrayProductos'));
-                arrayProductos.push(data);
-                localStorage.setItem('arrayProductos', JSON.stringify(arrayProductos));
-                console.log(arrayProductos);
-            });
-            
+        let idProduct = localStorage.getItem("IdProduct"); // Corregir si es necesario
+        if (idProduct) {
+            let URL = "https://japceibal.github.io/emercado-api/products/" + idProduct + ".json";
+            fetch(URL)
+                .then(res => {
+                    if (!res.ok) {
+                        throw new Error('Error en la solicitud fetch');
+                    }
+                    return res.json();
+                })
+                .then(data => {
+                    let arrayProductos = JSON.parse(localStorage.getItem('arrayProductos')) || [];
+                    // Verificar si el producto ya está en el carrito 
+                    const productExists = arrayProductos.some(product => product.id === data.id);
+                    if (!productExists) {
+                        arrayProductos.push(data);
+                        localStorage.setItem('arrayProductos', JSON.stringify(arrayProductos));
+                        console.log(arrayProductos);
+                    } else {
+                        Swal.fire({
+                            title:"Ya se encuentra en el carrito"
+
+                          })
+                    }
+                })
+                .catch(error => {
+                    console.error(error);
+                });
+        } else {
+            console.error('IdProduct no encontrado en el almacenamiento local.');
+        }
     });
-})
+});
+
 
 function showProduct(product) {
     let htmlContentToAppend = "";
